Use reactive-banana for FRP
I would add reactive-banana-wx as a Cabal dependency, but it insists on
having cabal-macosx installed as a Setup.hs dependency. Since proper
Setup.hs dependency management hasn't made it into a Cabal release yet,
this would require a manual package installation step which I'd like to
avoid if possible. So, for now, I'm pulling it in as a submodule.

On the other hand, I may actually need cabal-macosx for the editor to
run on OS X. If so, I'll add it in as a required dependency and swap out
the submodule for a Cabal dependency on reactive-banana-wx.